medical-rag-bot/requirements.txt
streamlit==1.33.0
langchain==0.1.16
langchain-community==0.0.34
langchain-google-genai==1.0.3
chromadb==0.4.24
sentence-transformers==2.7.0
pypdf==4.2.0
python-dotenv==1.0.1
